+++ src/FolkerKinzel.Strings.Tests/StringExtensionTests.cs
@@ -70,5 +70,163 @@ namespace FolkerKinzel.Strings.Tests
         [TestMethod()]
         [ExpectedException(typeof(ArgumentException))]
         public void GetPersistentHashCodeTest7() => _ = string.Empty.GetPersistentHashCode((HashType)4711);
+
+
+
+        [TestMethod]
+        public void TrimTest1()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+
+            string test = "\"\'Test\'\"";
+
+            Assert.AreEqual(test.Trim(trimChars), test.Trim(trimChars.AsSpan()));
+        }
+
+        [TestMethod]
+        public void TrimTest2()
+        {
+            string test = "  Test  ";
+
+            Assert.AreEqual(test.Trim(null), test.Trim(ReadOnlySpan<char>.Empty));
+        }
+
+        [TestMethod]
+        public void TrimTest3()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+            string test = "\"\'\'\"";
+            Assert.AreEqual(test.Trim(trimChars), test.Trim(trimChars.AsSpan()));
+        }
+
+
+        [TestMethod]
+        public void TrimTest4()
+        {
+            string test = "    ";
+            Assert.AreEqual(test.Trim(null), test.Trim(ReadOnlySpan<char>.Empty));
+        }
+
+
+        [TestMethod]
+        [ExpectedException(typeof(ArgumentNullException))]
+        public void TrimTest5()
+        {
+            string? test = null;
+            _ = test!.Trim(ReadOnlySpan<char>.Empty);
+        }
+
+        [TestMethod]
+        public void TrimTest6()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+
+            string test = "Test";
+            Assert.AreEqual(test.Trim(trimChars), test.Trim(trimChars.AsSpan()));
+        }
+
+        [TestMethod]
+        public void TrimStartTest1()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+
+            string test = "\"\'Test\'\"";
+
+            Assert.AreEqual(test.TrimStart(trimChars), test.TrimStart(trimChars.AsSpan()));
+        }
+
+        [TestMethod]
+        public void TrimStartTest2()
+        {
+            string test = "  Test  ";
+
+            Assert.AreEqual(test.TrimStart(null), test.TrimStart(ReadOnlySpan<char>.Empty));
+        }
+
+        [TestMethod]
+        public void TrimStartTest3()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+            string test = "\"\'\'\"";
+            Assert.AreEqual(test.TrimStart(trimChars), test.TrimStart(trimChars.AsSpan()));
+        }
+
+
+        [TestMethod]
+        public void TrimStartTest4()
+        {
+            string test = "    ";
+            Assert.AreEqual(test.TrimStart(null), test.TrimStart(ReadOnlySpan<char>.Empty));
+        }
+
+
+        [TestMethod]
+        [ExpectedException(typeof(ArgumentNullException))]
+        public void TrimStartTest5()
+        {
+            string? test = null;
+            _ = test!.TrimStart(ReadOnlySpan<char>.Empty);
+        }
+
+        [TestMethod]
+        public void TrimStartTest6()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+
+            string test = "Test";
+            Assert.AreEqual(test.TrimStart(trimChars), test.TrimStart(trimChars.AsSpan()));
+        }
+        
+        [TestMethod]
+        public void TrimEndTest1()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+
+            string test = "\"\'Test\'\"";
+
+            Assert.AreEqual(test.TrimEnd(trimChars), test.TrimEnd(trimChars.AsSpan()));
+        }
+
+        [TestMethod]
+        public void TrimEndTest2()
+        {
+            string test = "  Test  ";
+
+            Assert.AreEqual(test.TrimEnd(null), test.TrimEnd(ReadOnlySpan<char>.Empty));
+        }
+
+        [TestMethod]
+        public void TrimEndTest3()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+            string test = "\"\'\'\"";
+            Assert.AreEqual(test.TrimEnd(trimChars), test.TrimEnd(trimChars.AsSpan()));
+        }
+
+
+        [TestMethod]
+        public void TrimEndTest4()
+        {
+            string test = "    ";
+            Assert.AreEqual(test.TrimEnd(null), test.TrimEnd(ReadOnlySpan<char>.Empty));
+        }
+
+
+        [TestMethod]
+        [ExpectedException(typeof(ArgumentNullException))]
+        public void TrimEndTest5()
+        {
+            string? test = null;
+            _ = test!.TrimEnd(ReadOnlySpan<char>.Empty);
+        }
+
+        [TestMethod]
+        public void TrimEndTest6()
+        {
+            char[] trimChars = new char[] { '\'', '\"' };
+
+            string test = "Test";
+            Assert.AreEqual(test.TrimEnd(trimChars), test.TrimEnd(trimChars.AsSpan()));
+        }
     }
 }